Updated Relationships and Controllers

+++ Menlyn_Mews_API/Models/Domain/Price.cs
@@ -6,7 +6,7 @@ namespace Menlyn_Mews_API.Models.Domain
     {
         [Key]
         public int PriceId { get; set; }
-        public int Product_Price {  get; set; } = int.MinValue;
+        public Decimal Product_Price {  get; set; } = int.MinValue;
         public DateTime Price_Date { get; set; } = DateTime.MinValue;
 
         public virtual ICollection<Product> Product { get; set; }
